replace and rename solution2. Update all locations previously depending on Solution and expecting vectors
