+++ dfusion_rust_core/src/models/util.rs
@@ -1,6 +1,7 @@
 use byteorder::{BigEndian, ByteOrder};
 use graph::data::store::{Entity, Value};
 use graph::prelude::BigInt;
+use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::convert::TryInto;
 use std::str::FromStr;
@@ -219,4 +220,8 @@ pub fn read_amount(bytes: &[u8; 12]) -> u128 {
         .collect::<Vec<u8>>();
 
     BigEndian::read_u128(bytes.as_slice())
+}
+
+pub fn map_from_list(arr: &[(u16, u128)]) -> HashMap<u16, u128> {
+    arr.iter().copied().collect()
 }
