+++ driver/src/driver/stablex_driver.rs
@@ -76,6 +76,7 @@ mod tests {
 
     use dfusion_core::models::account_state::test_util::*;
     use dfusion_core::models::order::test_util::create_order_for_test;
+    use dfusion_core::models::util::map_from_list;
 
     use mock_it::Matcher::{Any, Val};
 
@@ -109,7 +110,7 @@ mod tests {
             .will_return(Ok(()));
 
         let solution = Solution {
-            prices: vec![1, 2],
+            prices: map_from_list(&[(0, 1), (1, 2)]),
             executed_sell_amounts: vec![0, 2],
             executed_buy_amounts: vec![0, 2],
         };
@@ -151,7 +152,7 @@ mod tests {
             .will_return(Ok(()));
 
         let solution = Solution {
-            prices: vec![1, 2],
+            prices: map_from_list(&[(0, 1), (1, 2)]),
             executed_sell_amounts: vec![0, 2],
             executed_buy_amounts: vec![0, 2],
         };
